+++ app/src/androidTest/java/woowacourse/movie/presentation/view/navigation/ScreeningMovieFragmentTest.kt
@@ -13,6 +13,7 @@ import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
 import woowacourse.movie.R
+import woowacourse.movie.presentation.view.screening.ScreeningMovieFragment
 import woowacourse.movie.presentation.view.screening.adapter.MovieViewHolder
 
 @RunWith(AndroidJUnit4::class)
