+++ app/src/main/java/woowacourse/movie/presentation/view/reservation/detail/MovieDetailActivity.kt
@@ -8,7 +8,7 @@ import android.widget.ArrayAdapter
 import woowacourse.movie.R
 import woowacourse.movie.databinding.ActivityMovieDetailBinding
 import woowacourse.movie.presentation.base.BaseActivity
-import woowacourse.movie.presentation.view.navigation.ScreeningMovieFragment.Companion.MOVIE_ID_KEY
+import woowacourse.movie.presentation.view.screening.ScreeningMovieFragment.Companion.MOVIE_ID_KEY
 import woowacourse.movie.presentation.view.reservation.seat.SeatSelectionActivity
 import woowacourse.movie.presentation.view.screening.theater.TheaterBottomSheetDialogFragment.Companion.THEATER_ID_KEY
